+++ src/libical/sspm.c
@@ -442,7 +442,7 @@ char* sspm_lowercase(char* str)
     }
     new = sspm_strdup(str);
     for(p = new; *p!=0; p++){
-	*p = tolower(*p);
+	*p = tolower((int)*p);
     }
 
     return new;
@@ -1148,13 +1148,13 @@ char *decode_quoted_printable(char *dest,
 		continue;
 	    }
 
-	    cc  = isdigit(*src) ? (*src - '0') : (*src - 55);
+	    cc  = isdigit((int)*src) ? (*src - '0') : (*src - 55);
 	    cc *= 0x10;
 	    src++; 
 	    if (!*src) {
 		break;
 	    }
-	    cc += isdigit(*src) ? (*src - '0') : (*src - 55);
+	    cc += isdigit((int)*src) ? (*src - '0') : (*src - 55);
 
 	    *dest = cc;
